Move NullArtifact to the packer base classes to reuse it
