--- builder/null/artifact_export_test.go
@@ -1,10 +1,0 @@
-package null
-
-import (
-	"github.com/mitchellh/packer/packer"
-	"testing"
-)
-
-func TestNullArtifact(t *testing.T) {
-	var _ packer.Artifact = new(NullArtifact)
-}
