+++ builder/null/builder.go
@@ -59,7 +59,9 @@ func (b *Builder) Run(ui packer.Ui, hook packer.Hook, cache packer.Cache) (packe
 	}
 
 	// No errors, must've worked
-	artifact := &NullArtifact{}
+	artifact := &packer.NullArtifact{
+		BuilderIdValue: BuilderId,
+	}
 	return artifact, nil
 }
