+++ packer/null_artifact_test.go
@@ -1,0 +1,9 @@
+package packer
+
+import (
+	"testing"
+)
+
+func TestNullArtifact(t *testing.T) {
+	var _ Artifact = new(NullArtifact)
+}
